% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiomaRtAnnotation.R
\name{getBiomaRtAnnotation}
\alias{getBiomaRtAnnotation}
\title{Get gene annotation from biomaRt Ensembl database}
\usage{
getBiomaRtAnnotation(identifierList, identifierType, species)
}
\arguments{
\item{identifierList}{values of filters that are applied to biomaRt query}

\item{identifierType}{types of filters that are applied to biomaRt query}

\item{species}{should be specified as "hsapiens" or "mmusculus"}
}
\value{
Annotation retrieved from biomaRt database
}
\description{
Get gene annotation based on identifier (e.g. Ensembl gene id) using biomaRt Ensembl database
}
