% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverlap.R
\name{getOverlap}
\alias{getOverlap}
\title{Overlap between genomic and exon coordinates}
\usage{
getOverlap(grObject, exonAnnotation)
}
\arguments{
\item{grObject}{GRanges object with genomic coordinates (see GenomicRanges documentation for more info)}

\item{exonAnnotation}{annotation file filtered for exon coordinates (e.g. from GENCODE)}
}
\value{
Table listing all overlapping coordinates with their corresponding overlap type
}
\description{
Compares supplied coordinates with coordinates from an exon annotation file (e.g. GENCODE .gtf)
}
