% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftOverGRange.R
\name{liftOverGRange}
\alias{liftOverGRange}
\title{Wrapper function for UCSC's liftOver tool}
\usage{
liftOverGRange(grObject, givenAssembly, targetAssembly, libPath)
}
\arguments{
\item{grObject}{GRanges object with genomic coordinates (see GenomicRanges documentation for more info)}

\item{givenAssembly}{genome assembly of the supplied coordinates (e.g. "hg19")}

\item{targetAssembly}{genome assembly to which the coordinates will be converted}

\item{libPath}{directory where the liftOver chain is stored (the chain will be downloaded if not already present in this directory)}
}
\value{
GRanges object with genomic coordinates after liftOver conversion
}
\description{
Converts coordinates in genomic range object from the given assembly to the target assembly
}
