% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorrplot.R
\name{makeCorrplot}
\alias{makeCorrplot}
\title{Generating a correlation plot}
\usage{
makeCorrplot(x, y = NULL, method = "pearson", adjust = "fdr")
}
\arguments{
\item{x}{Matrix or dataframe containing only numeric data}

\item{y}{An optional second matrix or dataframe with the same number of rows as x}

\item{method}{Options are "pearson", "spearman" or "kendall"}

\item{adjust}{Method for p-value correction; Options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" and "none"}
}
\description{
Generating a correlation plot
}
