% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlapOverview.R
\name{plotOverlapOverview}
\alias{plotOverlapOverview}
\title{Plotting function for overlap between genomic and exon coordinates}
\usage{
plotOverlapOverview(overlap, plotPrefix, outDir)
}
\arguments{
\item{overlap}{output of the "getOverlap" function}

\item{plotPrefix}{prefix for plot filename}

\item{outDir}{output directory}
}
\value{
None
}
\description{
Creates two barplots for an overview of how supplied coordinates match to the exon annotation file
}
